run data

+++ src/start.py
@@ -27,39 +27,39 @@ def main():
     _start = datetime.now()
     if exists('../http.log'):
         os.remove('../http.log')
-    #if downloads_files():
-    logger.info(f'Tempo de Dowload Time:{datetime.now() - _start}')
-    logger.info('iniciando a limpesa do banco Banco linado')
-    #clear_tables()
-    logger.info('Banco limpo')
+    if downloads_files():
+        logger.info(f'Tempo de Dowload Time:{datetime.now() - _start}')
+        logger.info('iniciando a limpesa do banco Banco linado')
+        clear_tables()
+        logger.info('Banco limpo')
         # Create cempa_tifs
-    #    tifs_path = f'{settings.CATALOG}cempa_tifs'
-    #    if isdir(tifs_path):
-    #        rmtree(tifs_path)
-    #    mkdir(tifs_path)
-    #    color_bar = f'{settings.CATALOG}colorbar'
-    #    if isdir(color_bar):
-    #        rmtree(color_bar)
-    #    mkdir(color_bar)
-    #    meta_path = f'{settings.CATALOG}cempa_metadata'
-    #    if isdir(meta_path):
-    #        rmtree(meta_path)
-    #    mkdir(meta_path)
+        tifs_path = f'{settings.CATALOG}cempa_tifs'
+        if isdir(tifs_path):
+            rmtree(tifs_path)
+        mkdir(tifs_path)
+        color_bar = f'{settings.CATALOG}colorbar'
+        if isdir(color_bar):
+            rmtree(color_bar)
+        mkdir(color_bar)
+        meta_path = f'{settings.CATALOG}cempa_metadata'
+        if isdir(meta_path):
+            rmtree(meta_path)
+        mkdir(meta_path)
 
-    layer = to_db()
+        layer = to_db()
 
-    restart_ows = post(initial_config.CEMPA_OWS_URL)
-    if restart_ows.status_code == 204:
-        time.sleep(5)
-        if isdir(initial_config.OWS_CACH):
-            try:
-                rmtree(initial_config.OWS_CACH)
-            except Exception:
-                logger.exception('Error ao limpar o cach')
-        logger.log('CEMPA', f'ows reiniciado e cach limpo')
-        time.sleep(120)
-        logger.log('CEMPA', f'inicinado criacao do cach')
-        run(layer)
+        restart_ows = post(initial_config.CEMPA_OWS_URL)
+        if restart_ows.status_code == 204:
+            time.sleep(60)
+            if isdir(initial_config.OWS_CACH):
+                try:
+                    rmtree(initial_config.OWS_CACH)
+                except Exception:
+                    logger.exception('Error ao limpar o cach')
+            logger.log('CEMPA', f'ows reiniciado e cach limpo')
+            time.sleep(60)
+            logger.log('CEMPA', f'inicinado criacao do cach')
+            run(layer)
 
     # else:
     #    pass
